+++ src/nikita/coursework/composite/GVEOval.java
@@ -11,20 +11,20 @@ import java.awt.geom.Ellipse2D;
  */
 public class GVEOval extends GVEShape {
 
-    public GVEOval(int posX, int posY, int width, int height) {
+    public GVEOval(double posX, double posY, double width, double height) {
         super(posX, posY, width, height);
     }
 
     @Override
     public void draw(Graphics g) {
         Graphics2D g2d = (Graphics2D) g.create();
-        Ellipse2D oval = new Ellipse2D.Float(x, y, width, height);
+        Ellipse2D oval = new Ellipse2D.Double(x, y, width, height);
         g2d.draw(oval);
         g2d.dispose();
     }
 
     @Override
-    public void move(int x, int y) {
+    public void move(double x, double y) {
         //TODO: Implement
     }
 }
